==> notebooks/1 - preprocessing/10 - preprocessing_part1.ipynb <==
# Preprocessing

In this notebook we filter out finals games since we are only interested in modelling the Brownlow Medal
We remove columns that are not relevant to the overall analysis, and columns that contain far too many N/A values

import pandas as pd

data = pd.read_csv('../../data/landing/player_stats_12-22_fry.csv')

data

# filter out any finals games
# round numbers appear as strings in some games, and int in others,
# therfore we account for both:
round_numbers = [i for i in range(1, 24)] + [str(j) for j in range(1,24)]
data = data.query('match_round.isin(@round_numbers)')

data.columns

data.isna().sum().sort_values(ascending=False)[:25]

drop_list = [
    # features irrelevant to model:
    'Unnamed: 0', 'match_local_time', 'venue_name', 'match_attendance',
    'match_weather_temp_c', 'match_weather_type',
    'player_height_cm', 'player_weight_kg', 'player_is_retired','guernsey_number',
    'subbed', 'date',
    # too many n/a values, supercoach scores will be sourced from another dataset:
    'supercoach_score'
]

# can get effective disposals from disposals and effective percentage as neither contains n/a
data['effective_disposals'] = round(data['disposal_efficiency_percentage'] / 100 * data['disposals'])

df = data.drop(columns=drop_list)

df

df.to_csv('../../data/raw/stats_12-22.csv', index=False)
