==> notebooks/1 - preprocessing/11 - preprocessing_na_values.ipynb <==
import pandas as pd

df = pd.read_csv('../../data/raw/stats_12-22.csv')

df.head()

df = df.astype({'match_date': 'datetime64[ns]'})

df['season'] = df.match_date.dt.year

len(df[df.isnull().any(axis=1)])

# null values arise in these years:
df[df.isnull().any(axis=1)].season.unique()

# Investigating 2012-2016 values

It seems for whatever odd reason, there are 2 or 3 matches in each season where the advanced statistics were not recorded
I have searched the official AFL website as well as other common sources such as afl tables and footywire, but they all return n/a.

df_2012 = df.query('season == 2012')
df_2012_na = df_2012[df_2012.isna().any(axis=1)]

df_2012.isna().sum().sort_values(ascending=False)[:25]

# the n/a values arise in 2 matches: Adelaide, GWS in round 4 and GC, Carlton round 22
# it seems the values were not recorded as they are not present on the official AFL website, or any other data keeping site that I have seen

df_2012_na

# 2013

df_2013 = df.query('season == 2013')
df_2013_na = df_2013[df_2013.isna().any(axis=1)]

df_2013.isna().sum().sort_values(ascending=False)[:25]

df_2013_na

df_2014 = df.query('season == 2014')
df_2014_na = df_2014[df_2014.isna().any(axis=1)]

df_2014.isna().sum().sort_values(ascending=False)[:25]

df_2014_na

df_2015 = df.query('season == 2015')
df_2015_na = df_2015[df_2015.isna().any(axis=1)]

df_2015.isna().sum().sort_values(ascending=False)[:25]

df_2015_na

df_2016 = df.query('season == 2016')
df_2016_na = df_2016[df_2016.isna().any(axis=1)]

df_2016.isna().sum().sort_values(ascending=False)[:25]

df_2016_na.iloc[[0, 60, 130]]

# Investigating 2020-2021 values

df_2020 = df.query('season == 2020')
df_2020_na = df_2020[df_2020.isna().any(axis=1)]

df_2020.isna().sum().sort_values(ascending=False)[:2]

# afl fantasy scores were not recorded in dataset in all 9 round 1 matches
len(df_2020_na.match_id.unique())

idx_to_change = df_2020_na.index

# calculate fantasy score via formula on afl website
# https://www.fantasy.afl.com.au/classic/help/game-guidelines

df.loc[idx_to_change, 'afl_fantasy_score'] = 3*df['kicks'] + 2*df['handballs'] \
                                            + 3*df['marks'] +4*df['tackles'] \
                                            + 1*df['free_kicks_for'] \
                                            - 3*df['free_kicks_against'] \
                                            + df['hitouts'] \
                                            + 6*df['goals'] \
                                            + 1*df['behinds']

# visible values agree with AFL website
df.loc[idx_to_change][['player_first_name', 'player_last_name', 'afl_fantasy_score']]

# Editing 2021 values

df_2021 = df.query('season == 2021')

df_2021_na = df_2021[df_2021.isna().any(axis=1)]

# only 10 player's rating points are missing
df_2021.isna().sum().sort_values(ascending=False)[:2]

col = [
    'match_home_team', 'match_away_team', 'match_round',
    'player_first_name', 'player_last_name', 'rating_points', 'disposals'
]

# create table to easily find missing values on AFL website
df_2021_na[col]

# fortunately, only 10 errors means it is easy to source by hand
rating_points_in_order = [
    7.3,
    0.5,
    2.8,
    19.8,
    3.3,
    10.6,
    4.8,
    8.5,
    14.3,
    5.5
]

idx_to_change = df_2021_na.index

df.loc[idx_to_change, 'rating_points'] = rating_points_in_order

# check values entered as intended
df.loc[idx_to_change, ['player_last_name', 'rating_points']]

# Final Comments

To begin, there was 889 total rows of n/a values
After investigating the reason for the missing values, only 484 remain.

These 484 correspond with 11 games * 44 players where individual statistics were simply 
not recorded during the play. This means a total of 33 brownlow polling performances will not 
be learned by the model.

Fortunately the missing performances occured in 2012-1016 and therefore do not impact
the years 2021 and 2022, which will be used for testing the model.

df[df.isna().any(axis=1)]

df.dropna(inplace=True)

df.to_csv('../../data/raw/stats_12-22_no_na.csv', index=False)
