==> notebooks/1 - preprocessing/12 - get_player_info_12-22.ipynb <==
# Player Information

We use this notebook to get a dataframe of player names, id and team for each season and add a column that transforms the the player's name in such a way that it can be merged with dataframes from different sources.

This transformation will come in useful when we merge the historic supercoach data as the player_id's do not match, and also their names are represented in a different fashion. e.g. Thomas Stewart vs. Tom Stewart

import pandas as pd

df = pd.read_csv('../../data/raw/stats_12-22_no_na.csv')

df.columns

df = df[['match_id', 'season', 'player_id', 'player_first_name', 'player_last_name', 'player_team']]

df.drop_duplicates(['player_id', 'season'], inplace=True)

df.query('player_id == 11706')

def get_multi_team(row):
    """returns the number of teams a player has played for"""

    player_id = row['player_id']

    return len(df.query('player_id == @player_id').player_team.unique())

df['no_teams'] = df.apply(get_multi_team, axis=1)

df.sort_values('no_teams')

def transform_name(row):
    
    """
    function to transform player name such that we take the first name initial and last name
    in the case of the 4 players where this is not unique (when paired with player team) we take the full first name
    puncuation is removed from player's names as well

    e.g. 
    1. Lewis Young => lewis young
    2. Sam De Koning => s de koning
    3. Jaeger O'Meara => j omeara
    4. Jeremy Cameron => j cameron
    """

    # 4 unique player ids where intial surname appears twice on same team in same season
    # 12567: Lewis Young, WB,
    # 12717: Lachie Young, WB,
    # 11644: Brad Ebert, PA,
    # 11323: Brad Ebert, PA,
    # fortunately all names are simple: not requiring any punctuation removal
    # since these are the only cases where the shortened name and team match
    # we will use their full name instead
    
    if row['player_id'] in {12567, 12717, 11644, 11323}:

        first = row['player_first_name']
        last = row['player_last_name']
        return_name = f'{first.lower()} {last.lower()}'

        return return_name


    first = row['player_first_name']
    lst = row['player_last_name'].replace("'", "").split('-')


    if '-' in row['player_last_name']:
        return_name = f'{first[0]} {lst[0][0]}-{lst[1]}'
        return return_name.lower()
    
    else:
        return_name = f'{first[0]} {lst[0]}'

        return return_name.lower()

df['process_name'] = df.apply(transform_name, axis=1)

# created a list of attributes that act as a unique identifier
df.groupby(['season', 'process_name', 'player_team']).count().sort_values('no_teams')

out_df = df[['season', 'player_id', 'player_team', 'no_teams', 'player_first_name', 'player_last_name', 'process_name']]

out_df

out_df.to_parquet('../../data/curated/player_information_12-22.parquet')
