==> notebooks/1 - preprocessing/13 - preprocessing_part2.ipynb <==
# Preprocessing Part 2

In this notebook we configure the datatypes of each column to the most appropriate.
Dataframes are now saved in parquet file format to retain datatypes and for the space advantage

import pandas as pd

df = pd.read_csv('../../data/raw/stats_12-22_no_na.csv')
df.head()

df.dtypes

# most columns will be integer type
# this will make type casting easier next cell
for col in df.columns:
    print(f"'{col}':'int',")

df = df.astype({
    'match_id':'int',
    'match_home_team':'string',
    'match_away_team':'string',
    'match_date':'datetime64[ns]',
    'match_round':'int',
    'match_home_team_goals':'int',
    'match_home_team_behinds':'int',
    'match_home_team_score':'int',
    'match_away_team_goals':'int',
    'match_away_team_behinds':'int',
    'match_away_team_score':'int',
    'match_margin':'int',
    'match_winner':'string',
    'player_id':'int',
    'player_first_name':'string',
    'player_last_name':'string',
    'player_team':'string',
    'kicks':'int',
    'marks':'int',
    'handballs':'int',
    'disposals':'int',
    'effective_disposals':'int',
    'disposal_efficiency_percentage':'float',
    'goals':'int',
    'behinds':'int',
    'hitouts':'int',
    'tackles':'int',
    'rebounds':'int',
    'inside_fifties':'int',
    'clearances':'int',
    'clangers':'int',
    'free_kicks_for':'int',
    'free_kicks_against':'int',
    'brownlow_votes':'category',
    'contested_possessions':'int',
    'uncontested_possessions':'int',
    'contested_marks':'int',
    'marks_inside_fifty':'int',
    'one_percenters':'int',
    'bounces':'int',
    'goal_assists':'int',
    'time_on_ground_percentage':'int',
    'afl_fantasy_score':'int',
    'centre_clearances':'int',
    'stoppage_clearances':'int',
    'score_involvements':'int',
    'metres_gained':'int',
    'turnovers':'int',
    'intercepts':'int',
    'tackles_inside_fifty':'int',
    'contest_def_losses':'int',
    'contest_def_one_on_ones':'int',
    'contest_off_one_on_ones':'int',
    'contest_off_wins':'int',
    'def_half_pressure_acts':'int',
    'effective_kicks':'int',
    'f50_ground_ball_gets':'int',
    'ground_ball_gets':'int',
    'hitouts_to_advantage':'int',
    'hitout_win_percentage':'int',
    'intercept_marks':'int',
    'marks_on_lead':'int',
    'pressure_acts':'int',
    'rating_points':'int',
    'ruck_contests':'int',
    'score_launches':'int',
    'shots_at_goal':'int',
    'spoils':'int',
    'player_position':'category',
    'season':'int'
})

df.columns

reorder_cols = [
    'match_id', 'match_home_team', 'match_away_team', 'match_date',
       'match_round', 'season', 'match_home_team_goals', 'match_home_team_behinds',
       'match_home_team_score', 'match_away_team_goals',
       'match_away_team_behinds', 'match_away_team_score', 'match_margin',
       'match_winner', 'player_id', 'player_first_name', 'player_last_name',
       'player_team','player_position', 'kicks', 'marks', 'handballs', 'disposals',
       'effective_disposals', 'disposal_efficiency_percentage', 'goals',
       'behinds', 'hitouts', 'tackles', 'rebounds', 'inside_fifties',
       'clearances', 'clangers', 'free_kicks_for', 'free_kicks_against',
       'brownlow_votes', 'contested_possessions', 'uncontested_possessions',
       'contested_marks', 'marks_inside_fifty', 'one_percenters', 'bounces',
       'goal_assists', 'time_on_ground_percentage', 'afl_fantasy_score',
       'centre_clearances', 'stoppage_clearances', 'score_involvements',
       'metres_gained', 'turnovers', 'intercepts', 'tackles_inside_fifty',
       'contest_def_losses', 'contest_def_one_on_ones',
       'contest_off_one_on_ones', 'contest_off_wins', 'def_half_pressure_acts',
       'effective_kicks', 'f50_ground_ball_gets', 'ground_ball_gets',
       'hitouts_to_advantage', 'hitout_win_percentage', 'intercept_marks',
       'marks_on_lead', 'pressure_acts', 'rating_points', 'ruck_contests',
       'score_launches', 'shots_at_goal', 'spoils'
]

df = df[reorder_cols]

df.to_parquet('../../data/raw/cleaned_stats_12-22.parquet')
