==> notebooks/1 - preprocessing/14 - preprocessing_footytables.ipynb <==
## Preprocessing Footytables Data

In this notebook we preprocess the data collected from footy tables as it contains the brownlow votes from 2022.

The aim is to derive the columns of interest and merge it with our original dataframe so that we have the brownlow votes information for all seasons.

Unfortunately there is no nice key to merge the dataframes on and when merging via the first and last name of players, some the result was not as expected. This was because of the represenation of some names that included, two words like "De Goey" or "De Koning", names with apostrophes like "O'hallaron" and names with an initial like "Josh P. Kennedy". There is also the case where names such as "Thomas" were shortened to "Tom"

So the names had to transformed to the same format so that they could be merged as intended.

import pandas as pd

# load afl_tables dataset
df = pd.read_csv('../../data/landing/player_stats_22_tables.csv', index_col=0)

df.columns

df

rounds = [str(i) for i in range(1,24)]
df = df.query('Round.isin(@rounds)')

keep_col = [
    'Home.team', 'Away.team', 'Round', 'First.name', 'Surname', 'Playing.for', 'Brownlow.Votes'
]

df = df[keep_col]

for col in keep_col:
    print(f'"{col}": "{col.lower().replace(".","_")}",')

df.rename(columns={
    "Home.team": "home_team",
    "Away.team": "away_team",
    "Round": "round_number",
    "First.name": "first_name",
    "Surname": "surname",
    "Playing.for": "player_team",
    "Brownlow.Votes": "brownlow_votes"
}, inplace=True)

df = df.astype({
    'round_number':'int'
})

df

# realistically we only need to merge rows where a player polled, and all remaining will be na
# which can be filled with .fillna(0)
polled = df.query('brownlow_votes > 0')

polled.head()

# 3 players * 9 games * 22 rounds (exclude bye round)
len(polled)

players = pd.read_parquet('../../data/curated/player_information_12-22.parquet').query('season == 2022')

players

# process_name and player_team together is a unique identifier
len(players.drop_duplicates(['process_name', 'player_team']))

def transform_name(row):
    
    """
    function to transform player name such that we take the first name initial and last name
    in the case of the 4 players where this is not unique (when paired with player team) we take the full first name
    puncuation is removed from player's names as well

    e.g. 
    1. Sam De Koning => s de koning
    2. Jaeger O'Meara => j omeara
    3. Jeremy Cameron => j cameron
    """

    first = row['first_name']
    lst = row['surname'].replace("'", "").split('-')


    if '-' in row['surname']:
        return_name = f'{first[0]} {lst[0][0]}-{lst[1]}'
        return return_name.lower()
    
    else:
        return_name = f'{first[0]} {lst[0]}'

        return return_name.lower()

polled['process_name'] = polled.apply(transform_name, axis=1)

new_df = pd.merge(polled, players, 
                  
        left_on=['process_name', 'player_team'],
        right_on=['process_name', 'player_team'], 
        how='left')

# 1 name represented incorrectly

len(new_df[new_df.isna().any(axis=1)])

new_df[new_df.isna().any(axis=1)]

polled.query('surname.str.contains("Rioli")')

players.query('player_last_name.str.contains("Rioli")')

polled.loc[5489, ['first_name', 'process_name']] = 'Junior', 'j rioli'

# remerge datasets after altering player name
new_df = pd.merge(polled, players, 
                  
        left_on=['process_name', 'player_team'],
        right_on=['process_name', 'player_team'], 
        how='left')


len(new_df[new_df.isna().any(axis=1)])

# top 20 looks correct as compared to aftables website
new_df.groupby(['player_id', 'player_first_name', 'player_last_name'])['brownlow_votes'].sum().reset_index().sort_values('brownlow_votes', ascending=False)[:20]

new_df

new_df = new_df[['player_id', 'player_first_name', 'player_last_name', 'player_team', 'round_number', 'brownlow_votes']]

new_df.drop(columns=['player_first_name', 'player_last_name', 'player_team'],inplace=True)

full_df = pd.read_csv('../../data/raw/stats_12-22_no_na.csv')

df_2022 = full_df.query('season == 2022').drop(columns='brownlow_votes')

df_2022 = pd.merge(df_2022, new_df,
         left_on=['player_id', 'match_round'],
         right_on=['player_id', 'round_number'],
         how='left').drop(columns='round_number')

# total players - players polled = n/a
# 22*9*46 - 22*9*3 = 8514 as expected
df_2022.isna().sum().sort_values(ascending=False)[:5]

df_2022.fillna(0, inplace=True)

df_2022.to_parquet('../../data/curated/brownlow_votes_22.parquet')

full_df = full_df.query('season < 2022')
full_df

final_df = pd.concat([full_df, df_2022])

final_df

# all columns merged succesfully and columns are same between both dataframes
final_df.isna().sum()

final_df.to_parquet('../../data/raw/cleaned_stats_12-22_fixed_bv.parquet')
