==> notebooks/1 - preprocessing/15 - preprocessing_coaches_votes.ipynb <==
# Preprocessing Coaches Votes

In this notebook we clean up the data that contains coaches votes from each game.
This includes mapping the game and player to their respective match_id and player_id's so they can be joined 
onto a future dataframe with a primary key

import pandas as pd

data = pd.read_csv('../../data/landing/coaches_votes_12-22.csv', index_col=0).reset_index(drop=True)

data

# first things first, let's clean up those column names
for col in list(data.columns):
    
    col_out = col.replace('.','_').lower()
    
    print(f"'{col}': '{col_out}',")

# add prefix "match_" for future df merges
data = data.rename(columns={
    'Season': 'season',
    'Round': 'match_round',
    'Home.Team': 'match_home_team',
    'Away.Team': 'match_away_team',
    'Player.Name': 'player_name',
    'Coaches.Votes': 'coaches_votes'
})

# also filter out finals games
match_round = [i for i in range(1, 24)]
data = data.query('match_round.isin(@match_round)')

# much nicer
data.head()

match_id = pd.read_parquet('../../data/raw/cleaned_stats_12-22_fixed_bv.parquet')

match_id = match_id[['match_id', 'match_home_team', 'match_away_team', 
                     'match_round', 'season', 'match_date', 
                     'player_id', 'player_first_name', 'player_last_name']]

match_id.head()

old_teams = data.match_home_team.unique()
new_teams = match_id.match_home_team.unique()
# teams where name is different
list(set(old_teams) - set(new_teams))

list(set(new_teams) - set(old_teams))

# extract player name from player_name column (i.e. without team)
data['player'] = data['player_name'].str.extract(r"(.*?) \(", expand=False)
# extract player team from player_name column
data['player_team'] = data['player_name'].str.extract(r"\((.*?)\)", expand=False)

data.player_team.unique()

new_teams

team_dict = {
    # coaches_vote format: match_id format
    'GWS Giants':'Greater Western Sydney',
    'Sydney Swans':'Sydney',
    'West Coast Eagles':'West Coast',
    'Gold Coast Suns':'Gold Coast',
    'Geelong Cats':'Geelong',
    'Adelaide Crows':'Adelaide',
    
    # shortened team names: long team names
    'SYD': 'Sydney', 
    'GWS': 'Greater Western Sydney', 
    'CARL': 'Carlton', 
    'RICH': 'Richmond', 
    'COLL': 'Collingwood', 
    'HAW': 'Hawthorn', 
    'BL': 'Brisbane Lions', 
    'MELB': 'Melbourne', 
    'ADEL': 'Adelaide',    
    'GCFC': 'Gold Coast', 
    'FRE': 'Fremantle', 
    'GEEL': 'Geelong', 
    'ESS': 'Essendon', 
    'NMFC': 'North Melbourne', 
    'WCE': 'West Coast', 
    'WB': 'Western Bulldogs', 
    'PORT': 'Port Adelaide', 
    'STK': 'St Kilda'
}

data.replace({'match_home_team':team_dict, 'match_away_team':team_dict, 'player_team':team_dict}, inplace=True)

data.head()

match_id['player'] = match_id['player_first_name'] + ' ' + match_id['player_last_name']

match_id.head()

df = pd.merge(match_id, data, how='left', left_on=['match_home_team', 'match_away_team', 'match_round', 'season', 'player'],
                                    right_on=['match_home_team', 'match_away_team', 'match_round', 'season', 'player'])

df

cols_keep = ['match_id', 'player_id', 'coaches_votes']
final_df = df[cols_keep].fillna(0)

final_df

final_df.to_parquet('../../data/raw/coaches_votes.parquet')
