==> notebooks/1 - preprocessing/21 - preprocessing_afl_part1.ipynb <==
# Preprocessing 2023 AFL data

Cleaning the data sourced from the offical AFL website

We rename the columns and remove any duplicate/unimportant columns.

import pandas as pd

df = pd.read_csv('../../data/landing/player_stats_23_afl.csv', index_col=0)

df

df.columns

for col in df.columns:
    print(f"'{col}': '{col.lower().replace('.','_')}',")

df = df.rename(columns={
    'providerId': 'provider_id',
    'utcStartTime': 'utcstarttime',
    'status': 'status',
    'compSeason.shortName': 'compseason_shortname',
    'round.name': 'round_name',
    'round.roundNumber': 'round_number',
    'venue.name': 'venue_name',
    'home.team.name': 'home_team_name',
    'home.team.club.name': 'home_team_club_name',
    'away.team.name': 'away_team_name',
    'away.team.club.name': 'away_team_club_name',
    'player.jumperNumber': 'player_jumpernumber',
    'player.photoURL': 'player_photourl',
    'player.player.position': 'player_position',
    'player.player.player.playerId': 'player_id',
    'player.player.player.captain': 'player_captain',
    'player.player.player.playerJumperNumber': 'playerjumpernumber',
    'player.player.player.givenName': 'player_first_name',
    'player.player.player.surname': 'player_last_name',
    'teamId': 'teamid',
    'gamesPlayed': 'gamesplayed',
    'timeOnGroundPercentage': 'timeongroundpercentage',
    'goals': 'goals',
    'behinds': 'behinds',
    'superGoals': 'supergoals',
    'kicks': 'kicks',
    'handballs': 'handballs',
    'disposals': 'disposals',
    'marks': 'marks',
    'bounces': 'bounces',
    'tackles': 'tackles',
    'contestedPossessions': 'contested_possessions',
    'uncontestedPossessions': 'uncontested_possessions',
    'totalPossessions': 'total_possessions',
    'inside50s': 'inside_fifties',
    'marksInside50': 'marks_inside_fifty',
    'contestedMarks': 'contested_marks',
    'hitouts': 'hitouts',
    'onePercenters': 'one_percenters',
    'disposalEfficiency': 'disposal_efficiency',
    'clangers': 'clangers',
    'freesFor': 'free_kicks_for',
    'freesAgainst': 'free_kicks_against',
    'dreamTeamPoints': 'afl_fantasy_score',
    'rebound50s': 'rebounds',
    'goalAssists': 'goal_assists',
    'goalAccuracy': 'goal_accuracy',
    'ratingPoints': 'rating_points',
    'ranking': 'ranking',
    'lastUpdated': 'lastupdated',
    'turnovers': 'turnovers',
    'intercepts': 'intercepts',
    'tacklesInside50': 'tackles_inside_fifty',
    'shotsAtGoal': 'shots_at_goal',
    'goalEfficiency': 'goalefficiency',
    'shotEfficiency': 'shotefficiency',
    'interchangeCounts': 'interchangecounts',
    'scoreInvolvements': 'score_involvements',
    'metresGained': 'metres_gained',
    'clearances.centreClearances': 'centre_clearances',
    'clearances.stoppageClearances': 'stoppage_clearances',
    'clearances.totalClearances': 'clearances',
    'extendedStats.effectiveKicks': 'effective_kicks',
    'extendedStats.kickEfficiency': 'kick_efficiency',
    'extendedStats.kickToHandballRatio': 'kicktohandballratio',
    'extendedStats.effectiveDisposals': 'effective_disposals',
    'extendedStats.marksOnLead': 'marks_on_lead',
    'extendedStats.interceptMarks': 'intercept_marks',
    'extendedStats.contestedPossessionRate': 'contested_possessionrate',
    'extendedStats.hitoutsToAdvantage': 'hitouts_to_advantage',
    'extendedStats.hitoutWinPercentage': 'hitout_win_percentage',
    'extendedStats.hitoutToAdvantageRate': 'hitout_to_advantage_rate',
    'extendedStats.groundBallGets': 'ground_ball_gets',
    'extendedStats.f50GroundBallGets': 'f50_ground_ball_gets',
    'extendedStats.scoreLaunches': 'score_launches',
    'extendedStats.pressureActs': 'pressure_acts',
    'extendedStats.defHalfPressureActs': 'def_half_pressure_acts',
    'extendedStats.spoils': 'spoils',
    'extendedStats.ruckContests': 'ruckcontests',
    'extendedStats.contestDefOneOnOnes': 'contest_def_one_on_ones',
    'extendedStats.contestDefLosses': 'contest_def_losses',
    'extendedStats.contestDefLossPercentage': 'contestdeflosspercentage',
    'extendedStats.contestOffOneOnOnes': 'contest_off_one_on_ones',
    'extendedStats.contestOffWins': 'contest_off_wins',
    'extendedStats.contestOffWinsPercentage': 'contest_off_wins_percentage',
    'extendedStats.centreBounceAttendances': 'centre_bounce_attendances',
    'extendedStats.kickins': 'kickins',
    'extendedStats.kickinsPlayon': 'kickinsplayon',
    'player.playerId': 'playerid',
    
    'player.captain': 'is_captain',

    'player.playerJumperNumber': 'playerjumpernumber',
    'player.givenName': 'player_givenname',
    'player.surname': 'player_surname',
    'teamStatus': 'teamstatus',
    'team.name': 'team_name'
})

df.columns

# remove finals matches
df = df.query('round_name.str.contains("Round")')

# remove duplicate columns and anything not relevant
cols_keep = [
    'round_number', 'venue_name',
    'home_team_name', 'away_team_name',
    'playerid', 'player_position',
    'player_first_name', 'player_last_name', 'teamid', 'team_name', 

    'timeongroundpercentage', 
    
    # player stats
    'goals', 'behinds', 'kicks',
    'handballs', 'disposals', 'marks', 'bounces', 'tackles',
    'contested_possessions', 'uncontested_possessions',
    'inside_fifties', 'marks_inside_fifty', 'contested_marks', 'hitouts',
    'one_percenters', 'disposal_efficiency', 'clangers', 'free_kicks_for',
    'free_kicks_against', 'afl_fantasy_score', 'rebounds', 'goal_assists',
    'rating_points', 'turnovers',
    'intercepts', 'tackles_inside_fifty',
     'score_involvements',
    'metres_gained', 'centre_clearances', 'stoppage_clearances',
    'clearances', 'effective_kicks', 'effective_disposals', 'marks_on_lead',
    'intercept_marks', 'hitouts_to_advantage',
    'hitout_win_percentage', 'ground_ball_gets',
    'f50_ground_ball_gets', 'score_launches', 'pressure_acts',
    'def_half_pressure_acts', 'spoils', 
    'contest_def_one_on_ones', 'contest_def_losses',
    'contest_off_one_on_ones',
    'contest_off_wins'
]

df = df[cols_keep]

matches = pd.DataFrame(df.groupby(['round_number', 'home_team_name']).sum()).sort_values(['round_number', 'home_team_name'])
match_id = [i for i in range(10_000, 10_000 + 23*9)]
matches['match_id'] = match_id

matches = matches['match_id']

df = pd.merge(df, matches, on=['round_number', 'home_team_name'])

df.head()

matches = df.drop_duplicates('match_id')[['match_id', 'round_number', 'home_team_name', 'away_team_name']].reset_index(drop=True)
matches.to_parquet('../../data/raw/2023_match_data.parquet')

df.columns

reorder_cols = [
    'match_id', 'round_number', 'venue_name', 'home_team_name', 'away_team_name',
       'playerid', 'player_first_name', 'player_last_name', 'player_position',
       'teamid', 'team_name', 'timeongroundpercentage', 
       
       'kicks', 'marks', 'handballs', 'disposals',
       'effective_disposals', 'goals',
       'behinds', 'hitouts', 'tackles', 'rebounds', 'inside_fifties',
       'clearances', 'clangers', 'free_kicks_for', 'free_kicks_against',
       'contested_possessions', 'uncontested_possessions',
       'contested_marks', 'marks_inside_fifty', 'one_percenters', 'bounces',
       'goal_assists', 'afl_fantasy_score',
       'centre_clearances', 'stoppage_clearances', 'score_involvements',
       'metres_gained', 'turnovers', 'intercepts', 'tackles_inside_fifty',
       'contest_def_losses', 'contest_def_one_on_ones',
       'contest_off_one_on_ones', 'contest_off_wins', 'def_half_pressure_acts',
       'effective_kicks', 'f50_ground_ball_gets', 'ground_ball_gets',
       'hitouts_to_advantage', 'hitout_win_percentage', 'intercept_marks',
       'marks_on_lead', 'pressure_acts', 'rating_points',
       'score_launches', 'spoils'
]

df = df[reorder_cols]

df.to_parquet('../../data/raw/stats_23.parquet')
