==> notebooks/1 - preprocessing/22 - get_player_details_23.ipynb <==
import pandas as pd

df = pd.read_parquet('../../data/raw/stats_23.parquet')

df = df.drop_duplicates('playerid')

df.columns

df = df[['playerid', 'player_first_name', 'player_last_name', 'team_name']]

df.to_parquet('../../data/raw/player_information_23.parquet')
