==> notebooks/1 - preprocessing/23 - preprocess_matches_2023.ipynb <==
import pandas as pd

df = pd.read_csv('../../data/landing/match_results_23.csv', index_col=0)

df.columns

df = df[['round.roundNumber',
    'home.team.name', 'home.score.goals', 'home.score.behinds', 'home.score.totalScore',
    'away.team.name', 'away.score.goals', 'away.score.behinds', 'away.score.totalScore',
    ]]

df.rename(columns={
    'round.roundNumber': 'round_number',
    'home.team.name': 'home_team', 
    'home.score.goals': 'home_team_goals', 
    'home.score.behinds': 'home_team_behinds', 
    'home.score.totalScore': 'home_team_total',
    'away.team.name': 'away_team', 
    'away.score.goals': 'away_team_goals', 
    'away.score.behinds': 'away_team_behinds', 
    'away.score.totalScore': 'away_team_total',
}, inplace=True)

df = df.query('round_number <= 24')

matches = pd.read_parquet('../../data/raw/2023_match_data.parquet')

matches

df = pd.merge(df, matches,
         left_on=['round_number', 'home_team' , 'away_team'],
         right_on=['round_number', 'home_team_name', 'away_team_name'],
         )

df.columns

df = df[[
    'match_id',
    'round_number', 
    'home_team', 'home_team_goals', 'home_team_behinds', 'home_team_total',
    'away_team', 'away_team_goals', 'away_team_behinds', 'away_team_total'
]]

df['match_margin'] = abs(df['home_team_total'] - df['away_team_total'])

def get_winner(row):

    if row['home_team_total'] > row['away_team_total']:
        return row['home_team']
    elif row['home_team_total'] < row['away_team_total']:
        return row['away_team']
    else:
        return 'Draw'

df['match_winner'] = df.apply(get_winner, axis=1)

df

df.to_parquet('../../data/raw/match_results.parquet')
