==> notebooks/1 - preprocessing/24 - preprocess_captains.ipynb <==
import pandas as pd

df = pd.read_csv('../../data/landing/is_player_captain_23.csv', index_col=0)

df

df['captain_list'] = df['Captain(s)'].str.split(', ').apply(lambda x: x if isinstance(x, list) else [x])

expanded = df['captain_list'].apply(lambda x: pd.Series(x)).fillna('na')

expanded.columns = [f'captain_{i+1}' for i in range(expanded.shape[1])]

df = df.join(expanded)[['Club', 'season', 'captain_1', 'captain_2', 'captain_3']]

df

melted_df = df.melt(
    id_vars=['Club', 'season'],
    value_vars=['captain_1', 'captain_2', 'captain_3'],
    value_name='name'
).drop(columns=['variable'])

melted_df = melted_df.query('name != "na"')

melted_df

players = pd.read_parquet('../../data/raw/player_information_23.parquet')

players['name'] = players['player_first_name'] + ' ' + players['player_last_name']

final_df = pd.merge(melted_df, players,
         on='name',
         how='left')

final_df

final_df['is_captain'] = 1

final_df.drop(columns=['name', 'Club', 'season'], inplace=True)

final_df

final_df.to_parquet('../../data/curated/captains_23.parquet')
