==> notebooks/1 - preprocessing/25 - preprocess_wheelo.ipynb <==
# Preprocessing Data from Wheeloratings.com

Fortunately within wheelo ratings' brownlow prediction csv they included coaches votes and supercoach score, so this is the source for these data points

unfortunately the download button is not an actual link, so I could not automate the download.

https://www.wheeloratings.com/afl_brownlow.html

import pandas as pd

# wheeloratings have a csv with uptodate coaches votes announced on awards night
df = pd.read_csv('../../data/landing/wheelo-brownlow-predictions.csv')

df.head()

df.columns

df = df[['Round', 'Player', 'CoachesVotes', 'Supercoach']]

df.head()

for col in df.columns:
    print(f'"{col}": "{col.lower()}",')

df = df.rename(columns={
    "Round": "round_number",
    "Player": "player",
    "CoachesVotes": "coaches_votes",
    "Supercoach": "SC"
})

df

# all values filled
df.isna().sum().sort_values(ascending=False)

# change name so can be merged
df.loc[df['player'].str.contains("Willie"), 'player'] = "Junior Rioli"

players = pd.read_parquet('../../data/raw/player_information_23.parquet')

players['player'] = players['player_first_name'] + ' ' + players['player_last_name']

players.query('player.str.contains("Rioli")')

df = pd.merge(df, players,
         on='player',
         how='left')

# merge successfully
df.loc[df.isna().any(axis=1)]

df = df[['round_number', 'player', 'playerid', 'team_name', 'coaches_votes', 'SC']]

df

df.to_parquet('../../data/raw/2023_coaches_SC.parquet')
