==> notebooks/1 - preprocessing/26 - merge_data.ipynb <==
import pandas as pd

df = pd.read_parquet('../../data/raw/stats_23.parquet')

df

# Match Results

matches = pd.read_parquet('../../data/raw/match_results.parquet')

matches

df = pd.merge(df, matches,
         how='left')

df

# Supercoach Scores and Coaches Votes

c_votes = pd.read_parquet('../../data/raw/2023_coaches_SC.parquet')

c_votes

df = pd.merge(df, c_votes).drop(columns='player')

# Captains Data

captains = pd.read_parquet('../../data/curated/captains_23.parquet')

captains

df = pd.merge(df, captains, how='left')

df.loc[df.is_captain.isna(), 'is_captain'] = 0

df

df.isna().sum().sort_values()

df.to_parquet('../../data/raw/total_23_data.parquet')
