==> notebooks/1 - preprocessing/27 - preprocessing_afl_part2.ipynb <==
import pandas as pd

df = pd.read_parquet('../../data/raw/total_23_data.parquet')

df

df.columns

df.isna().sum().sort_values(ascending=False)[:5]

# Feature Engineering

counted_stats = [
    'kicks', 'marks', 'handballs', 'disposals',
    'effective_disposals', 'goals', 'afl_fantasy_score', 'SC',
    'behinds', 'hitouts', 'tackles', 'rebounds', 'inside_fifties',
    'clearances', 'clangers', 'free_kicks_for', 'free_kicks_against',
    'contested_possessions', 'uncontested_possessions',
    'contested_marks', 'marks_inside_fifty', 'one_percenters', 'bounces',
    'goal_assists',
    'centre_clearances', 'stoppage_clearances', 'score_involvements',
    'metres_gained', 'turnovers', 'intercepts', 'tackles_inside_fifty',
    'contest_def_losses', 'contest_def_one_on_ones',
    'contest_off_one_on_ones', 'contest_off_wins', 'def_half_pressure_acts',
    'effective_kicks', 'f50_ground_ball_gets', 'ground_ball_gets',
    'hitouts_to_advantage', 'hitout_win_percentage', 'intercept_marks',
    'marks_on_lead', 'score_launches', 'spoils'
]

counted_stat_totals = df.groupby('match_id')[counted_stats].sum().add_suffix('_total')

counted_stat_totals

s = counted_stat_totals.eq(0).any()
drop_stat_count = list(s[s].index)
drop_stat_count

df = pd.merge(df, counted_stat_totals, on='match_id', how='inner')

df

for stat in counted_stats:
    # do not want to divide by 0
    if stat+'_total' not in drop_stat_count:
        df[stat+'_proportion'] = df[stat] / df[stat+'_total']

# Previous Polling Performance

prev_brownlow = pd.read_parquet('../../data/curated/brownlow_votes_22.parquet')

# length matches that of dropping duplicates by player_id
# therefore no 2 players have the same full name

# cannot merge via player_id since ids for fryzigg source is different from AFL website
# new players in 2023 do not have assigned fryzigg player_ids

len(prev_brownlow.drop_duplicates('player_id')) == len(prev_brownlow.drop_duplicates(['player_first_name', 'player_last_name']))


prev_brownlow

# Cripps: 29 votes across 21 games => 1.38 average
# looks good
prev = prev_brownlow.groupby(['player_first_name', 'player_last_name'])['brownlow_votes'].mean().sort_values(ascending=False)
prev

df = pd.merge(df, prev, 
                  on=['player_first_name', 'player_last_name'],
                  how='left')

df

# all game stats played by new players will have na
df.isna().sum().sort_values(ascending=False)[:4]

df.fillna(0, inplace=True)

df = df.rename(columns={'brownlow_votes': 'average_votes_prev'})

# Dummy Variables

# adds feature that equals 1 if a player obtains 30+ disposals and 2+ goals
df['30_and_2'] = ((df['disposals'] >= 30) & (df['goals'] >= 2)).astype(int)

# adds feature that equals 1 if a player manages 5 or more goals
df['high_goal_scorer'] = (df['goals'] >= 5).astype(int)

# if winning_margin does not predict well, include dummy variable whether the team won or not
# in the event of a draw, will return 0
df['won_match'] = (df['team_name'] == df['match_winner']).astype(int)

max_game_disposals = df.groupby('match_id')['disposals'].max().reset_index()
max_game_disposals.rename(columns={'disposals': 'game_high_disposals'}, inplace=True)

df = pd.merge(df, max_game_disposals, on='match_id', how='left')

df['top_disposal_getter'] = (df['disposals'] == df['game_high_disposals']).astype(int)

max_game_clearances = df.groupby('match_id')['clearances'].max().reset_index()
max_game_clearances.rename(columns={'clearances': 'game_high_clearances'}, inplace=True)

df = pd.merge(df, max_game_clearances, on='match_id', how='left')

df['top_clearance_getter'] = (df['clearances'] == df['game_high_clearances']).astype(int)

# Captains

captains = pd.read_parquet('../../data/curated/captains_23.parquet')

captains

df = pd.merge(df, captains, how='left')

# Player Positions

convert_pos = {
    'CHB': 'key_defender',
    'RR': 'rover',
    'HFFL': 'forward',
    'HBFR': 'defender',
    'HBFL': 'defender',
    'C': 'rover',
    'FB': 'key_defender',
    'CHF': 'key_forward',
    'WL': 'wing',
    'INT': 'INT',
    'FPL': 'forward',
    'R': 'rover',
    'WR': 'wing',
    'RK': 'ruck',
    'BPR': 'defender',
    'FF': 'key_forward',
    'SUB': 'sub',
    'HFFR': 'forward',
    'FPR': 'forward',
    'BPL': 'defender'
}
df.replace({'player_position': convert_pos}, inplace=True)

df = pd.get_dummies(df, columns=['player_position'], dtype=int)

df['midfielder_goals'] = df['goals'] * (df['player_position_rover'])

df

# Winning Margin

def get_winning_margin(row):

    if row['team_name'] == row['match_winner']:
        return int(row['match_margin'])
    
    else:
        return int(-1 * row['match_margin'])

df['winning_margin'] = df.apply(get_winning_margin, axis=1)

df.columns[:50]

df.columns[50:100]

df.columns[115:]

keep_cols = [
    'match_id', 'round_number', 'winning_margin', 
    'playerid', 'player_first_name', 'player_last_name', 'team_name', 'is_captain',

    'timeongroundpercentage',

    # stat proportions
    'kicks_proportion', 'marks_proportion', 'handballs_proportion',
    'disposals_proportion', 'effective_disposals_proportion',
    'goals_proportion', 'afl_fantasy_score_proportion', 'SC_proportion',
    'behinds_proportion', 'hitouts_proportion', 'tackles_proportion',
    'rebounds_proportion', 'inside_fifties_proportion',
    'clearances_proportion', 'clangers_proportion',
    'free_kicks_for_proportion', 'free_kicks_against_proportion',
    'contested_possessions_proportion',
    'uncontested_possessions_proportion', 'contested_marks_proportion',
    'marks_inside_fifty_proportion', 'one_percenters_proportion',
    'bounces_proportion', 'goal_assists_proportion',
    'centre_clearances_proportion', 'stoppage_clearances_proportion',
    'score_involvements_proportion', 'metres_gained_proportion',
    'turnovers_proportion', 'intercepts_proportion',
    'tackles_inside_fifty_proportion', 'contest_def_one_on_ones_proportion',
    'contest_off_one_on_ones_proportion',
    'def_half_pressure_acts_proportion', 'effective_kicks_proportion',
    'f50_ground_ball_gets_proportion', 'ground_ball_gets_proportion',
    'hitouts_to_advantage_proportion', 'hitout_win_percentage_proportion',
    'intercept_marks_proportion', 'marks_on_lead_proportion',
    'score_launches_proportion', 'spoils_proportion', 

    # binary features
    '30_and_2', 'high_goal_scorer', 'won_match', 'game_high_disposals',
    'top_disposal_getter', 'game_high_clearances', 'top_clearance_getter',
    # player position
    'player_position_INT', 'player_position_defender',
    'player_position_forward', 'player_position_key_defender',
    'player_position_key_forward', 'player_position_rover',
    'player_position_ruck', 'player_position_sub', 'player_position_wing',

    'midfielder_goals',


    # votes
    'coaches_votes', 'average_votes_prev'
]

df = df[keep_cols]
df.head()

df.to_parquet('../../data/curated/stats_2023_final.parquet')
